+++ frontend/src/components/@commons/Role.tsx
@@ -4,7 +4,7 @@ import interviewboth from 'assets/images/interviewboth.webp';
 import intervieweeImage from 'assets/images/interviewee.webp';
 import InterviewerImage from 'assets/images/interviewer.webp';
 
-import Image from 'components/@commons/Image';
+import Image from 'components/@commons/image/Image';
 
 const Role = ({ role }: RoleProp) => {
   return (
