+++ frontend/src/components/@commons/image/Image.tsx
