+++ frontend/src/components/interviewQuestion/InterviewQuestionSearchForm.tsx
@@ -5,9 +5,9 @@ import useInterviewQuestionSearch from 'hooks/questionSearch/useInterviewQuestio
 import deleteIcon from 'assets/images/close.svg';
 import searchIcon from 'assets/images/search.svg';
 
-import Image from 'components/@commons/Image';
 import Input from 'components/@commons/Input';
 import Button from 'components/@commons/button/Button';
+import Image from 'components/@commons/image/Image';
 
 const InterviewQuestionSearchForm = () => {
   const {
