+++ frontend/src/components/teams/AddMember.tsx
@@ -2,8 +2,8 @@ import styled from 'styled-components';
 
 import { GITHUB_AVATAR_SIZE_LIST } from 'constants/constants';
 
-import Image from 'components/@commons/Image';
 import Button from 'components/@commons/button/Button';
+import Image from 'components/@commons/image/Image';
 import { MemberType } from 'types/member';
 
 const AddMember = ({ addMember, removeEvent }: ParticipantProps) => {
