+++ frontend/src/pages/status/index.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import Image from 'components/@commons/Image';
 import Button from 'components/@commons/button/Button';
+import Image from 'components/@commons/image/Image';
 
 const ExceptionContainer = ({ children }: ExceptionContainerProps) => {
   return <S.Container>{children}</S.Container>;
